tests/e2e/requirements.txt
boto3
aws-cdk-lib
pytest
pytest-xdist
